+++ public/dialog_cursor_style.js
@@ -4,9 +4,10 @@ import {Res} from '/res.js';
 
 export class DialogCursorStyle extends DialogSelect {
 
-    constructor(core) {
-        super(core);
-        this.initialValue = this.core.getOption('cursorStyle');
+    constructor(core, initialValue) {
+        super();
+        this.core = core;
+        this.initialValue = initialValue;
     }
 
     onChange(event) {