+++ public/dialog_new_line_mode.js
@@ -4,9 +4,9 @@ import {Res} from '/res.js';
 
 export class DialogNewLineMode extends DialogSelect {
 
-    constructor(core) {
-        super(core);
-        this.initialValue = this.core.getOption('newLineMode');
+    constructor(initialValue) {
+        super();
+        this.initialValue = initialValue;
     }
 
     onChange(event) {