+++ public/dialog_print_margin_column.js
@@ -4,9 +4,10 @@ import {Res} from '/res.js';
 
 export class DialogPrintMarginColumn extends DialogPrompt {
 
-    constructor(core) {
-        super(core);
-        this.initialValue = this.core.getOption('printMarginColumn');
+    constructor(core, initialValue) {
+        super();
+        this.core = core;
+        this.initialValue = initialValue;
     }
 
     onChange(event) {