+++ public/dialog_wrap.js
@@ -4,9 +4,9 @@ import {Res} from '/res.js';
 
 export class DialogWrap extends DialogSelect {
 
-    constructor(core) {
-        super(core);
-        this.initialValue = this.core.getOption('wrap');
+    constructor(initialValue) {
+        super();
+        this.initialValue = initialValue;
     }
 
     onChange(event) {